[Refactor] remove some override iceberg classes
