+++ fe/fe-core/src/test/java/com/starrocks/server/CatalogLevelTest.java
@@ -24,7 +24,6 @@ import com.starrocks.catalog.Type;
 import com.starrocks.connector.hive.HiveMetaClient;
 import com.starrocks.connector.hive.HiveMetastoreApiConverter;
 import com.starrocks.connector.hive.HiveMetastoreTest;
-import com.starrocks.connector.iceberg.hive.HiveTableOperations;
 import com.starrocks.privilege.PrivilegeActions;
 import com.starrocks.qe.DDLStmtExecutor;
 import com.starrocks.sql.analyzer.AnalyzeTestUtil;
@@ -38,6 +37,7 @@ import mockit.Mocked;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.api.Table;
+import org.apache.iceberg.hive.HiveTableOperations;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
